+++ packages/@valapi/valorant-api.com/CHANGELOG.md
@@ -1,3 +1,13 @@
+# 3.2.0
+
+**Add**
+
+-   `ValorantApiCom.request()`
+
+**Remove**
+
+-   `ValorantApiCom.getService()`
+
 # 3.0.0
 
 **Add**